Adding multiple runs as well as some comparisons

+++ server/src/main/groovy/com/devshawn/coloring/server/entity/HeuristicResult.groovy
@@ -4,6 +4,6 @@ import com.devshawn.coloring.library.ColoringHeuristic
 
 class HeuristicResult {
     ColoringHeuristic heuristic
-    int colors
+    double colors
     long time
 }
